+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## v0.0.2
 
-* [#4](https://github.com/brodieG/recsub/issues/4): Substitution now respects
+* [#5](https://github.com/brodieG/recsub/issues/5): Substitution now respects
   search path order.
 
 ## v0.0.1